refactored container name resolution
